ReflectionObjec,ReflectionClass,allowFunction
Mock ReflectionObject and ReflectionClass to make them compatible with symfony 3.4 installation: the reason is that the method "getFileName" is used many time in symfony's code and its return result is the path to the file in the cache. This causes symfony can't find its config,translation,views,etc folders.
Make it possible to disable all functions except ones which were allowed because symfony 3.4 installation fails to run with errors of callbacks.